feature: 2.0.4-release: 1.去除对jedis的依赖，直接使用redisson 2.增加key前缀命名空间的设计 3.解决springApplicationContext注入问题

+++ dynamic-limiter-core/src/main/java/com/lingfengx/mid/dynamic/limiter/DLimiter.java
@@ -16,7 +16,7 @@ public class DLimiter {
     //需要限流的数量
     private int boxLen;
     //限流的时间窗口
-    private int boxTime;
+    private long boxTime;
     //是否组要限流
     private boolean limited;
 }
